requirements.txt
albumentations==1.3.0
map_boxes==1.0.5
matplotlib==3.4.1
nltk==3.8
numpy==1.23.5
opencv_python==4.7.0.68
opencv_python_headless==4.7.0.72
pandas==1.5.3
pandas==1.2.4
Pillow==9.4.0
Pillow==10.0.1
pycocotools==2.0.6
scikit_learn==1.3.0
skimage==0.0
tensorboardX==2.5.1
timm==0.6.13
torch==1.13.1
torchvision==0.14.1
tqdm==4.60.0
tqdm==4.65.0
transformers==4.30.2
